[CHORE] 타입 수정 | Modify type TALEARNT-391
