+++ src/pages/articles/MatchingArticleList/core/matchingArticleList.type.ts
@@ -1,8 +1,5 @@
-import {
-  durationType,
-  exchangeType,
-  matchingArticleBodyType
-} from "@pages/articles/core/articles.type";
+import { durationType, exchangeType } from "@pages/articles/core/articles.type";
+import { matchingArticleBodyType } from "@pages/articles/WriteArticle/core/writeArticle.type";
 import { profileType } from "@type/user.type";
 
 export type matchingArticleType = Pick<profileType, "nickname" | "profileImg"> &
