+++ src/pages/articles/core/articles.type.ts
@@ -13,14 +13,6 @@ export type durationType =
   | "3개월 이상";
 // 매칭 게시글 - 진행 방식
 export type exchangeType = "온라인" | "오프라인" | "온/오프라인";
-// 매칭 게시글 - body
-export type matchingArticleBodyType = commonArticleDataType & {
-  duration: durationType;
-  exchangeType: exchangeType;
-  giveTalents: number[];
-  receiveTalents: number[];
-  imageUrls: string[];
-};
 
 // 커뮤니티 게시글 - 게시판 타입
 export type postType = "자유 게시판" | "질문 게시판" | "스터디 모집";
